[Bronze III] Title: 이게 분수?, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/2863. 이게 분수？/이게 분수？.cc
@@ -2,9 +2,9 @@
 #include <cstdio>
 
 int main() {
-  double a, b, c, d;
-  scanf("%lf%lf%lf%lf", &a, &b, &c, &d);
+  float a, b, c, d;
+  scanf("%f%f%f%f", &a, &b, &c, &d);
 
-  double arr[4] = {a / c + b / d, c / d + a / b, d / b + c / a, b / a + d / c};
+  float arr[4] = {a / c + b / d, c / d + a / b, d / b + c / a, b / a + d / c};
   printf("%ld", std::max_element(arr, arr + 4) - arr);
 }
